claude/hooks/pyproject.toml
[project]
name = "hooks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
]
